% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{culture}
\alias{culture}
\title{Exponential growth of bacteria}
\format{A dataframe with 3 A600 measurements per time point:
 \describe{
 \item{min}{Units of time, minutes}
 \item{A600}{optical density at A600 nm}
}}
\source{
Murphy lab
}
\usage{
culture
}
\description{
Exponential growth of bacteria
}
\examples{
\dontrun{
 culture
}
}
\keyword{datasets}
