% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{decay1ex}
\alias{decay1ex}
\title{One phase decay example data}
\format{
A dataframe with 5 independent replicates:
 \describe{
 \item{Time}{Units of time, x.}
 \item{Response}{Response units, y.}
 \item{Replicate}{name of independent replicate.}
}
}
\source{
simulated with simdecay1
}
\usage{
decay1ex
}
\description{
One phase decay example data
}
\examples{
\dontrun{
 decay1ex
}
}
\keyword{datasets}
