% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decay2exdata.R
\docType{data}
\name{decay2ex}
\alias{decay2ex}
\title{Two phase decay example data}
\format{
A dataframe with 5 independent replicates:
 \describe{
 \item{Time}{Units of time, x.}
 \item{Response}{Response units, y.}
 \item{Replicate}{name of independent replicate.}
}
}
\source{
simdecay2 example
}
\usage{
decay2ex
}
\description{
Two phase decay example data
}
\examples{
\dontrun{
 decay2ex
}
}
\keyword{datasets}
