% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitggro.R
\name{fitggro}
\alias{fitggro}
\title{Fit Gompertz model for growth curves}
\usage{
fitggro(x, y, data, k, ylo, d, lambda, weigh)
}
\arguments{
\item{x}{A vector of linear values, usually time.}

\item{y}{A vector of measured values in response to x.}

\item{data}{A dataframe with x and y.}

\item{k}{The estimate for maximal growth rate.}

\item{ylo}{The estimate for the lowest value of y.}

\item{d}{The estimate for the difference between the
lowest and highest values of y.}

\item{lambda}{The estimate for the growth lag, in units of the x scale.}

\item{weigh}{When TRUE, weighted regression by 1/y^2. Default is FALSE.}
}
\value{
nls
}
\description{
Fit the Gompertz growth model to date:
`y =  ylo + d*exp(-exp(((k*exp(1))/d)*(lambda-x)+1))`
}
\examples{

# L.plantarium growth at 35C.

fitggro(x=hour, y=lnN_N0, data=plantdat,
k=1, ylo=0, d=8, lambda=3, weigh=FALSE)


}
