% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitlindr.R
\name{fitlindr}
\alias{fitlindr}
\title{Fit dose response when dose is linear scale.
Based on a model}
\usage{
fitlindr(x, y, data, k, ylo, yhi, h, weigh)
}
\arguments{
\item{x}{is a vector of linear scale values, usually dose or concentration.}

\item{y}{is a vector of y values for x, usually responses.}

\item{data}{is a dataframe with x and y.}

\item{k}{the value of x that yields y/ymax = 0.5, usually ED50 or EC50.}

\item{ylo}{is the lowest asymptotic y value, in response units.}

\item{yhi}{the highest asymptotic value, in response units.}

\item{h}{the Hill slope, a unitless slope factor; -1 > h > 1 is steeper, -1 < h < 1 is shallower. Use negative value for downward sloping response.}

\item{weigh}{chooses regression weighting by 1/y^2. Default = FALSE.}
}
\value{
model
}
\description{
Fit dose response when dose is linear scale.
Based on a model
}
\examples{
# fitlinder(x=PE, y=mN, data=aoc, k=100, ylo=0, yhi=50, h=1, weigh=T)

}
