% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitlogdr.R
\name{fitlogdr}
\alias{fitlogdr}
\title{Fit dose-response when dose is log10 scale}
\usage{
fitlogdr(x, y, data, logk, ylo, yhi, h, weigh)
}
\arguments{
\item{x}{a vector of base 10 log scale values, usually dose or concentration.}

\item{y}{a vector of y values for x, usually responses.}

\item{data}{a dataframe with x and y.}

\item{logk}{estimates the value of base 10 log scale x that yields y/ymax = 0.5, usually ED50 or EC50.}

\item{ylo}{estimates the lowest asymptotic y value, in response units.}

\item{yhi}{estimates highest asymptotic value, in response units.}

\item{h}{the Hill slope, a unitless slope factor; -1 > h > 1 is steeper, -1 < h < 1 is shallower. Use negative value for downward sloping response.}

\item{weigh}{chooses regression weighting by 1/y^2. Default = FALSE.}
}
\value{
nls
}
\description{
fitlogdr uses the nlsLM function of the minpack.lm package to fit a model
formula to nonlinear response data. The formula is based upon the general
hyperbolic function y= ymax/(1+10^((logk-x)*h)), where ymax = yhi - ylo.
Both x and logk are in base 10 log scale units. logk, ylo, yhi and h are
start list parameters #' for nlsLM.
}
\examples{
# aorta contraction by phenylephrine doses
fitlogdr(x=logM_PE, y=mN, data=aoc, logk=-8, ylo=0, yhi=50, h=1, weigh=TRUE)

# or
fitlogdr(x=lognM_PE, y=mN, data=aoc, logk=1, ylo=0, yhi=50, h=1, weigh=TRUE)



}
