% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plantdat}
\alias{plantdat}
\title{Gompertz model for growth curves}
\format{A dataframe:
\describe{
\item{hour}{Units of time, hour}
\item{lnN_N0}{l.plantarium growth measure}
}}
\source{
simulated with simggro
}
\usage{
plantdat
}
\description{
Gompertz model for growth curves
}
\examples{
\dontrun{
 plantdat
}
}
\keyword{datasets}
