% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdecay1.R
\name{simdecay1}
\alias{simdecay1}
\title{Simulate one phase exponential decay data}
\usage{
simdecay1(x, k, ylo, yhi, cv, reps)
}
\arguments{
\item{x}{a vector of non-exponential linear scale values representing time.}

\item{k}{the  rate constant, expressed in reciprocal of the X axis time units.
The half-life is 0.6932/k.}

\item{ylo}{the lowest expected y value, or the value at infinite times,
expressed in the same units as Y.}

\item{yhi}{the highest expected y value, or the starting value,
expressed in the same units as Y.}

\item{cv}{the coefficient of variation for y replicates.}

\item{reps}{an integer value for number of replicates}
}
\value{
ggplot, data
}
\description{
A sandbox to simulate and visualize random normal heteroscedastic response data.
 Variances enlarge with the value of y predicted by the model using a constant
 coefficeint of variation (cv). The data generating formula is derived from the
 general model: "y = y0*e^-kx". A one-phase model is used when It is used whenever
 the rate at which something happens is proportional to the amount which is left.
  Failure errors in the plot  fitting subfunction will occasionally happen due
  to the random data. These are more frequent with higher cv values. J
  ust re-simulate with modified parameter values.
}
\examples{

# Note: exponential or log-transformed x scale values will not work
# do not use x = c(1e-9, 3e-9, ...) or c(-9, -8.523, ...)

time <- c(1, 5, 10, 15, 20, 25) # eg, in mins

set.seed(2345)

decay1dat <- simdecay1(time, k = 0.15, ylo = 1, yhi = 100, cv = 0.10, reps = 5)


decay1dat$data


}
