% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simegro.R
\name{simegro}
\alias{simegro}
\title{Simulate exponential growth}
\usage{
simegro(x, k, ylo, sd, reps)
}
\arguments{
\item{x}{A vector of linear scale values, such as time..}

\item{k}{The  rate constant, expressed in reciprocal of the X axis units.
When time, the half-life is 0.6932/k.}

\item{ylo}{The lowest value of y, or a starting value.}

\item{sd}{The standard deviation of y.}

\item{reps}{An integer value for number of replicates.}
}
\value{
ggplot, data
}
\description{
A sandbox to simulate exponential growth with random normal error.
Enter a linear scale predictor variable, x, with arguments for rate
constant, k, and initial starting value of response, ylo. Include an
estimate for the standard deviation, sd, in y.  Random values for y are
simulated by: #' \code{y = ylo*exp(k*x) + rnorm(length(x), 0, sd)}.
The regression formula is: `y ~ ylo\emph{exp(k}x)``
}
\examples{
# x is time in minutes

simegro(x=c(1:10), k=0.3, ylo=10, sd=10, reps=3)


}
