% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simggro.R
\name{simggro}
\alias{simggro}
\title{Simulate growth curves}
\usage{
simggro(x, k, ylo, d, lambda, sd, reps)
}
\arguments{
\item{x}{A vector of linear scale values, such as time.}

\item{k}{The maximal growth rate, approximately the slope of the l
ine in the growth phase.}

\item{ylo}{The lowest value of y, or a starting value.}

\item{d}{The difference between ylo and the highest value of y.}

\item{lambda}{The lag period, in units of the x scale}

\item{sd}{The standard deviation of the replicates.}

\item{reps}{The number of replicates per value of x.}
}
\value{
ggplot, data
}
\description{
A sandbox to simulate a growth process, beginning with a
low initial lag period, through an exponential growth phase,
to a higher final maximal asymptote. The function is based
upon the Gompertz growth model. Random values for y are created using the equation:
\code{y =  ylo + d*exp(-exp(((k*exp(1))/d)*(lambda-x)+1)) + rnorm(length(x), 0, sd)}.
The regression formula is:
\code{y ~  ylo + d*exp(-exp(((k*exp(1))/d)*(lambda-x) + 1))}.
}
\examples{

# simulates l.plantarium growth, see Fig 3 in Applied Environ Micro 56:1875, 1990

plantarium <- simggro(x=c(1:8, 12, 13, 24),
  k=2, ylo=0, d=9, lambda=2.5, sd=0.5, reps=1)

plantarium

plantarium$data


}
