% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simhetdr.R
\name{simhetdr}
\alias{simhetdr}
\title{Simulate nonlinear heteroscedastic dose-response data}
\usage{
simhetdr(x, k, ylo, yhi, h, cv, reps, weight = F, log = F)
}
\arguments{
\item{x}{a vector of non-exponential linear scale values, usually representing dose or concentration, but can represent any stimulus.}

\item{k}{the value of x that yields y/ymax = 0.5, usually EC50 or ED50.}

\item{ylo}{the lowest expected y value, in response units.}

\item{yhi}{the highest expected y value, in response units.}

\item{h}{the Hill slope, a unitless slope factor; -1 > h > 1 is steeper, -1 < h < 1 is shallower. Use negative value for downward sloping response.}

\item{cv}{the coefficient of variation for y replicates.}

\item{reps}{an integer value for number of replicates}

\item{weight}{logical value indicating y scale weighting. Default is FALSE. If TRUE, curve is fit using relative (1/y^2) weighting.}

\item{log}{logical value. Default is FALSE. If TRUE, linear x values are transformed using a log10 function for plotting. Only for visual aesthetic.}
}
\value{
ggplot, data
}
\description{
A sandbox to simulate and visualize random normal heteroscedastic response data.
Variances enlarge with the value of y predicted by the model using a constant
coefficeint of variation (cv). The data generating formula is derived from the
general hyperbolic model: y/ymax=x^h/(x^h+k^h). Failure errors in the plot
fitting subfunction will occasionally happen due to the random data. These are
more frequent with higher cv values. Just re-simulate with modified parameter values.
The regression formula is \code{y ~ ylo + (yhi - ylo)*x^h/(x^h + k^h)}
}
\examples{

# Note: exponential or log-transformed x scale values will not work
# do not use x = c(1e-9, 3e-9, ...) or c(-9, -8.523, ...)

dose <- c(1, 3, 10, 30, 100, 300) # eg, in nM units

set.seed(2345)

hetdat <- simhetdr(dose, k = 35, ylo = 100, yhi = 1000,
h = 1.0, cv = 0.10, reps = 5, weight=TRUE, log = TRUE ); hetdat

hetdat$data


}
