% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simhetegro.R
\name{simhetegro}
\alias{simhetegro}
\title{Simulate heteroscedastic exponential growth data}
\usage{
simhetegro(x, k, ylo, cv, reps, weight = F)
}
\arguments{
\item{x}{A vector of linear scale values.}

\item{k}{A starting estimate for the rate constant, which has
units of reciprocal x.}

\item{ylo}{The lowest value of y, or starting value.}

\item{cv}{The coefficient of variation.}

\item{reps}{The number of replicates per value of x.}

\item{weight}{A logical value indicating y scale weighting. Default is FALSE.
If TRUE, the curve is fit using relative (1/y^2) weighting.}
}
\value{
ggplot, data
}
\description{
A sandbox to simulate heteroscedastic exponential growth data.
Random normal variance enlarges with the model value of y using
a constant coefficeint of variation (cv). The data generating
model is an exponential growth function: `yp=ylo*exp(k*x)`.
The simulated value is then `y=rnorm(1, yp, cv*yp)`.
The plot fitting subfunction will occasionally fail due to random
data. These may occur with higher cv values, fewer reps, and
fewer x values. Just re-simulate with modified parameter values.
}
\examples{

# Hypothetical Cell growth measured every 4 hours, replicated 5 times

simu <- simhetegro(x=c(0, 4, 8, 12, 16, 20, 24, 28), k=0.2, ylo=100,
cv=0.3, reps=5, weight=FALSE)

simu

simu$data#'

}
