% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmicmenten.R
\name{simmicmenten}
\alias{simmicmenten}
\title{Simulate Michaelis-Menten substrate vs. enzyme velocity data}
\usage{
simmicmenten(x, vmax, km, cv, reps, h = 1, log = F)
}
\arguments{
\item{x}{a vector of non-exponential linear scale values, usually representing dose or concentration, but can represent any stimulus.}

\item{vmax}{the maximum enzyme velocity possible. The curve approaches this value with increasing concentrations of substrate.}

\item{km}{the value of x that yields y/vmax = 0.5 (the Michaelis-Menten constant)}

\item{cv}{the coefficient of variation for y replicates}

\item{reps}{an integer value for number of replicates}

\item{h}{a value representing the hill slope coefficient. The default is h = 1 which considers enzyme-substrate interactions that only occur at 1 site on the enzyme. When h != 1, it indicates cooperativity (positive if h > 1; negative if h < 1). The h coefficient will not always equal the number of true binding sites, but it cannot exceed the number of true binding sites.}

\item{log}{logical value. Default is FALSE. If TRUE, linear x values are transformed using a log10 function for plotting. Only for visual aesthetic.}
}
\value{
ggplot, data
}
\description{
A sandbox to simulate and visualize random normal heteroscedastic response data. Variances enlarge with the value of y predicted by the model using a constant coefficeint of variation (cv). The data generating formula is derived from the Michaelis-Menten model equation including considerations for allosteric modulation: y=vmax*x^h/(x+km^h). Failure errors in the plot fitting subfunction will occasionally happen due to the random data. These are more frequent with higher cv values. Just re-simulate with modified parameter values.
}
\examples{

# Note: exponential or log-transformed x scale values will not work
# do not use x = c(1e-9, 3e-9, ...) or c(-9, -8.523, ...)

dose <- c(1, 3, 10, 30, 100, 300) # eg, in nM units

set.seed(2345)

vdat <- simmicmenten(dose, vmax = 1000, km = 50, cv = 0.10, reps = 5, h = 1, log = TRUE ); vdat

vdat$data

simmicmenten(dose, vmax = 10000, km = 100, cv = 0.25, reps = 10, h = 1.25, log = FALSE)


}
