% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aoc}
\alias{aoc}
\title{Contraction of mouse aorta in response to phenylephrine}
\format{A dataframe with 5 independent replicates:
\describe{
\item{nM_PE}{Doses of phenylephrine applied, in nM units.}
\item{replicate}{name of independent replicate.}
\item{mN}{measured contractile tension response in milliNewtons per mg dry weight.}
\item{M_PE}{Doses of phenylephrine applied, in M units.}
\item{logM_PE}{Doses of phenylephrine applied, in log10 M units.}
\item{lognM_PE}{Doses of phenylephrine applied, in log10 nM units.}
}}
\source{
Murphy lab
}
\usage{
aoc
}
\description{
Contraction of mouse aorta in response to phenylephrine
}
\examples{
\dontrun{
 aoc
}
}
\keyword{datasets}
