% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitlindr.R
\name{fitlindr}
\alias{fitlindr}
\title{Fit dose-response when dose is linear scale.}
\usage{
fitlindr(x, y, data, k, ylo, yhi, h, weigh)
}
\arguments{
\item{x}{a vector of linear scale values, usually dose or concentration.}

\item{y}{a vector of y values for x, usually responses.}

\item{data}{a dataframe with x and y.}

\item{k}{estimates the value of x that yields y/ymax = 0.5, usually ED50 or EC50.}

\item{ylo}{estimates the lowest asymptotic y value, in response units.}

\item{yhi}{estimates highest asymptotic value, in response units.}

\item{h}{the Hill slope, a unitless slope factor; -1 > h > 1 is steeper, -1 < h < 1 is shallower. Use negative value for downward sloping response.}

\item{weigh}{chooses regression weighting by 1/y^2. Default = FALSE.}
}
\value{
nls
}
\description{
fitlindr uses the nlsLM function from the minpack.lm package to fit a
model formula to nonlinear response data. The formula is based upon the
general hyperbolic function: y/ymax=x^h/(x^h+k^h), where ymax = yhi - ylo.
k, ylo, yhi and h are start list parameters for nlsLM.
}
\examples{
## aorta contraction by phenylephrine doses
fitlindr(x=nM_PE, y=mN, data=aoc, k=100, ylo=0, yhi=50, h=1, weigh=TRUE)

}
