% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim1sbind.R
\name{sim1sbind}
\alias{sim1sbind}
\title{Simulate one-site ligand receptor binding}
\usage{
sim1sbind(x, bmax, kd, cv, reps, log = F)
}
\arguments{
\item{x}{a vector of non-exponential linear scale values, usually representing dose or concentration of ligand}

\item{bmax}{the measured value where the receptor population is completely saturated by ligand (i.e. maximum binding)}

\item{kd}{the value of x that yields y/Bmax = 0.5 (the equilibrium binding constant)}

\item{cv}{the coefficient of variation for y replicates}

\item{reps}{an integer value for number of replicates}

\item{log}{logical value. Default is FALSE. If TRUE, linear x values are transformed using a log10 function for plotting. Only for visual aesthetic.}
}
\value{
ggplot, data
}
\description{
A sandbox to simulate and visualize random normal heteroscedastic response data. Variances enlarge with the value of y predicted by the model using a constant coefficeint of variation (cv). The data generating formula is derived from the one-site total binding model: y=Bmax*x/(x+kd). Failure errors in the plot fitting subfunction will occasionally happen due to the random data. These are more frequent with higher cv values. Just re-simulate with modified parameter values.
}
\examples{

# Note: exponential or log-transformed x scale values will not work
# do not use x = c(1e-9, 3e-9, ...) or c(-9, -8.523, ...)

dose <- c(1, 3, 10, 30, 100, 300) # eg, in nM units

set.seed(2345)

binddat <- sim1sbind(dose, bmax = 1000, kd = 50, cv = 0.10, reps = 5, log = FALSE ); binddat

binddat$data


}
