% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdecay1.R
\name{simdecay1}
\alias{simdecay1}
\title{Simulate one phase exponential decay data}
\usage{
simdecay1(x, k, ylo, yhi, sd, reps)
}
\arguments{
\item{x}{a vector of non-exponential linear scale values representing time.}

\item{k}{the  rate constant, expressed in reciprocal of the X axis time units.
The half-life is 0.6932/k.}

\item{ylo}{the lowest expected y value, or the value at infinite times,
expressed in the same units as Y.}

\item{yhi}{the highest expected y value, or the starting value,
expressed in the same units as Y.}

\item{sd}{the standard deviation in y.}

\item{reps}{an integer value for number of replicates}
}
\value{
ggplot, data
}
\description{
A sandbox to simulate and visualize random normal data
for a nonlinear decaying response. The data generating formula
is derived from the general model: "y = y0*e^-kx". This model
simulates response systems where the rate at which the response
decreases is proportional to the level of remaining response.
Failure errors can happen in the plot fitting subfunction even though random
data is produced. These may be more frequent with higher sd values and/or
lower range or replicates. Just re-simulate with modified parameter values.
The regression formula is: `y ~ (yhi-ylo)*exp(-1*k*x) + ylo`
}
\examples{

# Note: exponential or log-transformed x scale values will not work
# do not use x = c(1e-9, 3e-9, ...) or c(-9, -8.523, ...)

time <- c(1, 5, 10, 15, 20, 25) # eg, in mins

set.seed(2345)

decayOnedat <- simdecay1(time, k=0.15, ylo=1, yhi=100, sd = 10, reps=5)


decayOnedat$data


}
