% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simhetdr.R
\name{simhetdr}
\alias{simhetdr}
\title{Simulate nonlinear heteroscedastic dose-response data}
\usage{
simhetdr(x, k, ylo, yhi, h, sd, ex, reps, log = F)
}
\arguments{
\item{log}{}
}
\value{

}
\description{
A sandbox to simulate data
}
