% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simlindr.R
\name{simlindr}
\alias{simlindr}
\title{Simulate dose-response}
\usage{
simlindr(x, k, ylo, yhi, h, sd, reps)
}
\arguments{
\item{x}{A vector of linear scale values, usually dose or concentration..}

\item{k}{The value of x that yields y/ymax = 0.5, usually EC50 or ED50}

\item{ylo}{The lowest expected y value, in response units.}

\item{yhi}{The highest expected y value, in response units.}

\item{h}{The Hill slope, a unitless slope factor; -1 > h > 1 is steeper, -1 < h < 1 is shallower. Use negative value for downward sloping response.}

\item{sd}{Standard deviation of residual error, in response units.}

\item{reps}{Integer value for number of replicates.}
}
\value{
ggplot
}
\description{
Creates and ggplots replicate dose-response data with random normal error based on a linear scaled predictor variable and additional parameter arguments. Derived from the general hyperbolic function: y/ymax=logx^h/(logx^h+logk^h), where ymax = yhi - ylo.
}
\examples{
Up <- simlindr(x = c(1, 3, 30, 100, 300), k = 30,
ylo = 300, yhi = 3000,
h = 1.0,
sd = 100, reps = 5); Up

Up$data

Down <- simlindr(x = c(1, 3, 30, 100, 300), k = 30,
ylo = 300, yhi = 3000,
h = -1.0,
sd = 100, reps = 5); Down



}
